requirements.txt
playwright==1.52.0
pytest==8.4.2
allure-pytest==2.14.2
pydantic==2.11.9
pydantic-settings==2.11.0
